Replace obsolete project support component with OfferList
The commit deletes outdated classes related to project link handling in the user interface. This includes deletion of ProjectLink.java, ProjectLinksComponent.java, ProjectSupportComponent.java files. Instead of the generic ProjectLinksComponent.java an OfferList was introduced to ProjectInformationMain.java
