+++ user-interface/src/main/java/life/qbic/datamanager/views/projects/project/info/ProjectInformationMain.java
@@ -15,10 +15,12 @@ import com.vaadin.flow.spring.annotation.SpringComponent;
 import com.vaadin.flow.spring.annotation.UIScope;
 import jakarta.annotation.security.PermitAll;
 import java.io.Serial;
+import java.util.List;
 import life.qbic.application.commons.ApplicationException;
 import life.qbic.application.commons.Result;
 import life.qbic.datamanager.security.UserPermissions;
 import life.qbic.datamanager.views.Context;
+import life.qbic.datamanager.views.general.OfferDownload;
 import life.qbic.datamanager.views.notifications.StyledNotification;
 import life.qbic.datamanager.views.notifications.SuccessMessage;
 import life.qbic.datamanager.views.projects.project.ProjectMainLayout;
@@ -27,9 +29,18 @@ import life.qbic.datamanager.views.projects.project.experiments.ExperimentListCo
 import life.qbic.datamanager.views.projects.project.experiments.ExperimentListComponent.AddExperimentClickEvent;
 import life.qbic.datamanager.views.projects.project.experiments.ExperimentListComponent.ExperimentSelectionEvent;
 import life.qbic.datamanager.views.projects.project.experiments.experiment.create.AddExperimentDialog;
+import life.qbic.datamanager.views.projects.project.experiments.experiment.create.AddExperimentDialog.ExperimentAddEvent;
+import life.qbic.datamanager.views.projects.project.experiments.experiment.create.AddExperimentDialog.ExperimentDraft;
+import life.qbic.datamanager.views.projects.project.info.OfferList.DeleteOfferClickEvent;
+import life.qbic.datamanager.views.projects.project.info.OfferList.DownloadOfferClickEvent;
+import life.qbic.datamanager.views.projects.project.info.OfferList.OfferInfo;
+import life.qbic.datamanager.views.projects.project.info.OfferList.UploadOfferClickEvent;
+import life.qbic.datamanager.views.projects.purchase.UploadPurchaseDialog;
 import life.qbic.logging.api.Logger;
 import life.qbic.projectmanagement.application.AddExperimentToProjectService;
 import life.qbic.projectmanagement.application.OntologyTermInformationService;
+import life.qbic.projectmanagement.application.purchase.OfferDTO;
+import life.qbic.projectmanagement.application.purchase.ProjectPurchaseService;
 import life.qbic.projectmanagement.domain.model.experiment.ExperimentId;
 import life.qbic.projectmanagement.domain.model.project.Project;
 import life.qbic.projectmanagement.domain.model.project.ProjectId;
@@ -55,57 +66,46 @@ public class ProjectInformationMain extends Div implements BeforeEnterObserver,
   private static final Logger log = logger(ProjectInformationMain.class);
   private final transient AddExperimentToProjectService addExperimentToProjectService;
   private final transient OntologyTermInformationService ontologyTermInformationService;
-  private final UserPermissions userPermissions;
+  private final transient ProjectPurchaseService projectPurchaseService;
+  private final transient UserPermissions userPermissions;
   public static final String PROJECT_ID_ROUTE_PARAMETER = "projectId";
   public static final String EXPERIMENT_ID_ROUTE_PARAMETER = "experimentId";
   private final ProjectDetailsComponent projectDetailsComponent;
   private final ExperimentListComponent experimentListComponent;
-  private final ProjectLinksComponent projectLinksComponent; //TODO replace with OfferLinks
+  private final OfferDownload offerDownload;
+  private final OfferList offerList;
   private Context context;
 
   public ProjectInformationMain(@Autowired ProjectDetailsComponent projectDetailsComponent,
       @Autowired ExperimentListComponent experimentListComponent,
-      @Autowired ProjectLinksComponent projectLinksComponent,
       @Autowired UserPermissions userPermissions,
       @Autowired AddExperimentToProjectService addExperimentToProjectService,
-      @Autowired OntologyTermInformationService ontologyTermInformationService) {
+      @Autowired OntologyTermInformationService ontologyTermInformationService,
+      @Autowired ProjectPurchaseService projectPurchaseService) {
     this.projectDetailsComponent = requireNonNull(projectDetailsComponent,
         "projectDetailsComponent must not be null");
     this.experimentListComponent = requireNonNull(experimentListComponent,
         "experimentListComponent must not be null");
-    this.projectLinksComponent = requireNonNull(projectLinksComponent,
-        "projectLinksComponent must not be null");
     this.userPermissions = requireNonNull(userPermissions, "userPermissions must not be null");
     this.addExperimentToProjectService = requireNonNull(addExperimentToProjectService,
         "addExperimentToProjectService must not be null");
     this.ontologyTermInformationService = requireNonNull(ontologyTermInformationService,
         "ontologyTermInformationService must not be null");
+    this.projectPurchaseService = requireNonNull(projectPurchaseService,
+        "projectPurchaseService must not be null");
+
+    offerList = getConfiguredOfferList();
+    offerDownload = new OfferDownload(
+        (projectId, offerId) -> projectPurchaseService.getOfferWithContent(projectId, offerId)
+            .orElseThrow());
 
     this.experimentListComponent.addExperimentSelectionListener(this::onExperimentSelectionEvent);
     this.experimentListComponent.addAddButtonListener(this::onAddExperimentClicked);
 
     addClassNames("main", "project");
-    add(projectDetailsComponent, projectLinksComponent, experimentListComponent);
-
-    log.debug(String.format(
-        "New instance for %s(#%s) created with %s(#%s), %s(#%s) and %s(#%s)",
-        this.getClass().getSimpleName(), System.identityHashCode(this),
-        projectDetailsComponent.getClass().getSimpleName(),
-        System.identityHashCode(projectDetailsComponent),
-        experimentListComponent.getClass().getSimpleName(),
-        System.identityHashCode(experimentListComponent),
-        projectLinksComponent.getClass().getSimpleName(),
-        System.identityHashCode(projectLinksComponent)));
+    add(projectDetailsComponent, offerList, offerDownload, experimentListComponent);
   }
 
-  private void onAddExperimentClicked(AddExperimentClickEvent event) {
-    log.debug("Add experiment clicked: " + event);
-    showAddExperimentDialog();
-  }
-
-  private void onExperimentSelectionEvent(ExperimentSelectionEvent event) {
-    routeToExperiment(event.getExperimentId());
-  }
 
   /**
    * Extracts {@link ExperimentId} from the provided URL before the user accesses the page
@@ -130,14 +130,71 @@ public class ProjectInformationMain extends Div implements BeforeEnterObserver,
     }
   }
 
+  private OfferList getConfiguredOfferList() {
+    OfferList component = new OfferList();
+    component.addDeleteOfferClickListener(this::onDeleteOfferClicked);
+    component.addDownloadOfferClickListener(this::onDownloadOfferClicked);
+    component.addUploadOfferClickListener(
+        event -> onUploadOfferClicked(event, projectPurchaseService,
+            context.projectId().orElseThrow().value()));
+    return component;
+  }
+
+  private void onDownloadOfferClicked(DownloadOfferClickEvent downloadOfferClickEvent) {
+    offerDownload.trigger(context.projectId().orElseThrow().value(),
+        downloadOfferClickEvent.offerId());
+  }
+
+  private void onDeleteOfferClicked(DeleteOfferClickEvent deleteOfferClickEvent) {
+    projectPurchaseService.deleteOffer(context.projectId().orElseThrow().value(),
+        deleteOfferClickEvent.offerId());
+    deleteOfferClickEvent.getSource().remove(deleteOfferClickEvent.offerId());
+    offerDownload.removeHref();
+  }
+
+  private void onUploadOfferClicked(UploadOfferClickEvent uploadOfferClickEvent,
+      ProjectPurchaseService projectPurchaseService,
+      String projectId) {
+    UploadPurchaseDialog dialog = new UploadPurchaseDialog();
+    dialog.addConfirmListener(confirmEvent -> {
+      List<OfferDTO> offerDTOs = confirmEvent.getSource().purchaseItems().stream()
+          .map(it -> new OfferDTO(it.signed(), it.fileName(), it.content()))
+          .toList();
+      projectPurchaseService.addPurchases(projectId, offerDTOs);
+      refreshOffers(projectPurchaseService, projectId, uploadOfferClickEvent.getSource());
+      confirmEvent.getSource().close();
+    });
+    dialog.addCancelListener(cancelEvent -> cancelEvent.getSource().close());
+    dialog.open();
+  }
+
+  private static void refreshOffers(ProjectPurchaseService projectPurchaseService, String projectId,
+      OfferList offerList) {
+    List<OfferInfo> offers = projectPurchaseService.linkedOffers(projectId)
+        .stream()
+        .map(offer -> new OfferInfo(offer.id(), offer.getFileName(), offer.isSigned()))
+        .toList();
+    offerList.setOffers(offers);
+  }
+
+
+  private void onAddExperimentClicked(AddExperimentClickEvent event) {
+    log.debug("Add experiment clicked: " + event);
+    showAddExperimentDialog();
+  }
+
+  private void onExperimentSelectionEvent(ExperimentSelectionEvent event) {
+    routeToExperiment(event.getExperimentId());
+  }
+
   private void setContext(Context context) {
     this.context = context;
     projectDetailsComponent.setContext(context);
-    projectLinksComponent.setContext(context);
     experimentListComponent.setContext(context);
+    refreshOffers(projectPurchaseService, context.projectId().orElseThrow().value(), offerList);
   }
 
-  private void onExperimentAddEvent(AddExperimentDialog.ExperimentAddEvent event) {
+  private void onExperimentAddEvent(ExperimentAddEvent event) {
     ProjectId projectId = context.projectId().orElseThrow();
     ExperimentId createdExperiment = createExperiment(projectId, event.getExperimentDraft());
     event.getSource().close();
@@ -168,7 +225,7 @@ public class ProjectInformationMain extends Div implements BeforeEnterObserver,
   }
 
   private ExperimentId createExperiment(ProjectId projectId,
-      AddExperimentDialog.ExperimentDraft experimentDraft) {
+      ExperimentDraft experimentDraft) {
     Result<ExperimentId, RuntimeException> result = addExperimentToProjectService.addExperimentToProject(
         projectId,
         experimentDraft.getExperimentName(),
